refactor(web): move shared member/skeleton-overwrite UI out of pages/ (fix components->pages reach-ups)
